+++ example/gnu_make.mk
@@ -1,40 +1,32 @@
 # Makefile for Wirepas C Mesh API example, GNU make version
 
-# Toolchain
-CC  := gcc
-AS  := as
-LD  := ld
-AR  := ar
+# This example needs the C Mesh API library
+MESH_LIB_FOLDER := ../lib/
+MESH_LIB := $(MESH_LIB_FOLDER)build/mesh_api_lib.a
+
+# Detect platform and set toolchain variables
+include $(MESH_LIB_FOLDER)platform.mk
 
 # Paths, including trailing path separator
 SOURCEPREFIX   := ./
 BUILDPREFIX    := build/
-
-# This example needs the mesh lib
-MESH_LIB_FOLDER := ../lib/
-MESH_LIB := $(MESH_LIB_FOLDER)build/mesh_api_lib.a
-
-# General compiler flags
-CFLAGS  := -std=gnu99 -Wall -Werror
 
 # Targets definition
 MAIN_APP := meshAPIExample
 
 TARGET_APP := $(BUILDPREFIX)$(MAIN_APP)
 
-# Add Api header
+# Add API header
 CFLAGS  += -I$(MESH_LIB_FOLDER)api
-# Add pthtread lib as needed by Mesh Lib
-LDFLAGS += -pthread
-# Add Reentrant flag as using pthread
-CFLAGS  += -D_REENTRANT
 
-# Specific sources for main
+# Source files
 SOURCES:= $(SOURCEPREFIX)main.c
 
-OBJECTS := $(patsubst $(SOURCEPREFIX)%,                     \
-                  $(BUILDPREFIX)%,                          \
-                  $(SOURCES:.c=.o))
+# Object files
+OBJECTS := $(patsubst $(SOURCEPREFIX)%,         \
+                      $(BUILDPREFIX)%,          \
+                      $(SOURCES:.c=.o))
+
 
 # Functions
 
@@ -55,17 +47,14 @@ define CLEAN
 	$(RM) -r $(BUILDPREFIX)
 endef
 
+
 # Target rules
 
-# Use dependency files automatically generated by GCC.
-# First collect all C source files
-AUTODEPS := $(patsubst $(SOURCEPREFIX)%.c, $(BUILDPREFIX)%.d, $(SOURCES))
-
 ifeq ($(V),1)
-# "V=1" on command line, print commands.
+	# "V=1" on command line, print commands
 else
-# Default, do not print commands.
 .SILENT:
+	# Default, do not print commands
 endif
 
 .PHONY: all
