+++ lib/gnu_make.mk
@@ -1,29 +1,25 @@
 # Makefile for Wirepas C Mesh API library, GNU make version
 
-# Toolchain
-CC  := gcc
-AS  := as
-LD  := ld
-AR  := ar
+# Variables
+
+# Detect platform and set toolchain variables
+include platform.mk
 
 # Paths, including trailing path separator
-SOURCEPREFIX   :=
-BUILDPREFIX    := build/
+SOURCEPREFIX :=
+BUILDPREFIX  := build/
 
-# General compiler flags
-CFLAGS  := -std=gnu99 -Wall -Werror
 # Uncomment the following line only if running with an old stack
 #CFLAGS  += -DLEGACY_APP_CONFIG
 
 # Targets definition
 LIB_NAME := mesh_api_lib.a
-
 TARGET_LIB := $(BUILDPREFIX)$(LIB_NAME)
 
 SOURCES :=
 
-# Add Api header (including logger)
-CFLAGS  += -Iapi/
+# Add API header (including logger)
+CFLAGS += -Iapi/
 
 # Include platform module sources
 include platform/gnu_make.mk
@@ -34,6 +30,7 @@ include wpc/gnu_make.mk
 OBJECTS := $(patsubst $(SOURCEPREFIX)%,                     \
                       $(BUILDPREFIX)%,                      \
                       $(SOURCES:.c=.o))
+
 
 # Functions
 
@@ -47,7 +44,7 @@ endef
 
 define LINK_LIB
 	echo "  Linking $(1)"
-	$(AR) rc -o $(1) $(2)
+	$(ARCHIVE) $(1) $(2)
 endef
 
 define CLEAN
@@ -55,17 +52,14 @@ define CLEAN
 	$(RM) -r $(BUILDPREFIX)
 endef
 
+
 # Target rules
 
-# Use dependency files automatically generated by GCC.
-# First collect all C source files
-AUTODEPS := $(patsubst $(SOURCEPREFIX)%.c, $(BUILDPREFIX)%.d, $(SOURCES))
-
 ifeq ($(V),1)
-# "V=1" on command line, print commands.
+	# "V=1" on command line, print commands
 else
-# Default, do not print commands.
 .SILENT:
+	# Default, do not print commands
 endif
 
 .PHONY: all
