+++ lib/platform/posix/logger.c
