+++ lib/platform/posix/platform.c
