+++ test/gnu_make.mk
@@ -1,47 +1,40 @@
 # Makefile for Wirepas C Mesh API test suite, GNU make version
 
-# Toolchain
-CC  := gcc
-AS  := as
-LD  := ld
-AR  := ar
+# This example needs the C Mesh API library
+MESH_LIB_FOLDER := ../lib/
+MESH_LIB := $(MESH_LIB_FOLDER)build/mesh_api_lib.a
+
+# Detect platform and set toolchain variables
+include $(MESH_LIB_FOLDER)platform.mk
 
 # Paths, including trailing path separator
 SOURCEPREFIX   := ./
 BUILDPREFIX    := build/
-
-# This test suite needs the mesh lib
-MESH_LIB_FOLDER := ../lib/
-MESH_LIB := $(MESH_LIB_FOLDER)build/mesh_api_lib.a
-
-# General compiler flags
-CFLAGS  := -std=gnu99 -Wall -Werror
 
 # Targets definition
 MAIN_APP := meshAPItest
 
 TARGET_APP := $(BUILDPREFIX)$(MAIN_APP)
 
-# Add Api header
+# Add API header
 CFLAGS  += -I$(MESH_LIB_FOLDER)api
-# Add pthtread lib as needed by Mesh Lib
-LDFLAGS += -pthread
-# Add Reentrant flag as using pthread
-CFLAGS  += -D_REENTRANT
 
-# Test app needs some platform abstraction (It is a hack as not really part of lib API)
+# Test app needs some platform abstraction
+# This is a hack and not really part of the C Mesh API library public API
 CFLAGS  += -I$(MESH_LIB_FOLDER)platform
 
-# Main app
+# Main app source file
 SOURCES := $(SOURCEPREFIX)test_app.c
 
-# Test cases
+# Test case source files
 CFLAGS  += -I$(SOURCEPREFIX)
 SOURCES += $(SOURCEPREFIX)test.c
 
-OBJECTS := $(patsubst $(SOURCEPREFIX)%,                     \
-                  $(BUILDPREFIX)%,                          \
-                  $(SOURCES:.c=.o))
+# Object files
+OBJECTS := $(patsubst $(SOURCEPREFIX)%,         \
+                      $(BUILDPREFIX)%,          \
+                      $(SOURCES:.c=.o))
+
 
 # Functions
 
@@ -62,17 +55,14 @@ define CLEAN
 	$(RM) -r $(BUILDPREFIX)
 endef
 
+
 # Target rules
 
-# Use dependency files automatically generated by GCC.
-# First collect all C source files
-AUTODEPS := $(patsubst $(SOURCEPREFIX)%.c, $(BUILDPREFIX)%.d, $(SOURCES))
-
 ifeq ($(V),1)
-# "V=1" on command line, print commands.
+	# "V=1" on command line, print commands
 else
-# Default, do not print commands.
 .SILENT:
+	# Default, do not print commands
 endif
 
 .PHONY: all
